+++ src/withHooks/useLocalDataTrack/useLocalDataTrack.ts
@@ -1,0 +1,14 @@
+import { useMemo } from 'react';
+import { LocalDataTrack } from 'twilio-video';
+import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
+
+export function useLocalDataTrack() {
+  const { localTracks } = useVideoContext();
+
+  const localDT = useMemo(() => {
+    // It's OK to assume that there is a local data track.
+    return localTracks.find(track => track.kind === 'data') as LocalDataTrack;
+  }, [localTracks]);
+
+  return localDT;
+}